Added motor files, created gcode generator
